+++ packages/client/src/views/project/simple/images/Prompt.tsx
@@ -8,7 +8,7 @@ import { useFileInput } from '$hooks/useFileInput';
 import { useHotkey } from '$hooks/useHotkey';
 import { ProjectLoadPrompt } from '$modals/project';
 import { modalStore } from '$stores/ModalStore';
-import styles from './index.module.scss';
+import styles from './Prompt.module.scss';
 import { useSimpleProject } from '../../context';
 import { StyleSelect } from '../common/StyleSelect';